requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.0.0
scipy>=1.10.0
PyYAML>=6.0
openpyxl>=3.1.0
pyarrow>=12.0.0
